visualization/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
psycopg2-binary==2.9.7
python-dotenv==1.0.0
sqlalchemy>=2.0.30